pyproject.toml
[project]
name = "inventory-management-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fpdf2>=2.8.3",
    "mysql-connector-python>=9.3.0",
    "pandas>=2.3.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
]
